+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+
 ## [0.0.1] - 2020-03-02
 
 ### Added
@@ -14,3 +15,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added extra dirs: appcontext (create repository tool), config (environment variables), domain (sensu gateway structs and interfaces), gateway (sensu http gateway configurations), usecase (application logical interface between controller and gateway), utils(some utilities functions).
 - Added sensu-certs with json files to help to create all certificates using cfssl tools.
 - Added scripts for local test: run.sh, remove.sh, test-all.sh.
+- Added secret_env_from and secret_volume to mount previously created secrets in Sensu Agent.
+
+### Fixed
+- Sensu Handler Controller to add finalizer and update status.